--- app/stores/atom.ts
@@ -1,3 +1,0 @@
-import { atom } from 'jotai';
-
-export const fileAtom = atom<File | null>(null);
